moves the MapLayers to MapContainer
